refactor: linking transfer
adding navigation to transfer in Home and NavigationModal
