+++ package.json
@@ -14,6 +14,7 @@
     "@types/styled-components": "^5.1.9",
     "axios": "^0.21.1",
     "date-fns": "^2.22.1",
+    "polished": "^4.1.3",
     "react": "^17.0.2",
     "react-currency-input-field": "^3.3.2",
     "react-dom": "^17.0.2",
