+++ src/pages/Home/components/HeaderMainPage/styles.ts
@@ -1,8 +1,8 @@
-import React from 'react';
 import styled from 'styled-components';
-import { BiFilterAlt, BiSearch } from 'react-icons/bi';
 
-const Header = styled.header`
+import { darken } from 'polished';
+
+export const Header = styled.header`
   width: 100%;
   height: 200px;
   background-color: #fbaf00;
@@ -12,14 +12,14 @@ const Header = styled.header`
   position: relative;
 `;
 
-const SearchContainer = styled.div`
+export const SearchContainer = styled.div`
   width: calc(100% - 60px);
   display: flex;
   justify-content: space-between;
   margin-top: 40px;
 `;
 
-const SearchBox = styled.div`
+export const SearchBox = styled.div`
   width: calc(100% - 40px);
   height: 30px;
   background-color: white;
@@ -48,7 +48,7 @@ const SearchBox = styled.div`
   }
 `;
 
-const Icon = styled.div`
+export const Icon = styled.div`
   width: 30px;
   height: 30px;
   border-radius: 15px;
@@ -58,7 +58,7 @@ const Icon = styled.div`
   align-items: center;
 `;
 
-const Card = styled.div`
+export const Card = styled.div`
   height: 140px;
   width: calc(100% - 60px);
   background-color: white;
@@ -68,31 +68,38 @@ const Card = styled.div`
   border-radius: 10px;
   display: flex;
   align-items: center;
-  .extract {
+  > div {
     height: 85px;
     width: calc(100% - 140px);
     margin-left: 30px;
   }
-  .title {
-    font-size: 12px;
-    color: #6c6d71;
+
+  div {
+    > span {
+      font-size: 12px;
+      color: #6c6d71;
+    }
+
+    span + span {
+      font-weight: 500;
+      font-size: 18px;
+      line-height: 27px;
+    }
+
+    div {
+      width: 100%;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      margin-top: 20px;
+    }
   }
+
   hr {
     margin-top: 5px;
     margin-right: 50px;
   }
-  .money {
-    font-weight: 500;
-    font-size: 18px;
-    line-height: 27px;
-  }
-  .extract-container {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 20px;
-  }
+
   a {
     color: #e80234;
     font-weight: 500;
@@ -100,7 +107,9 @@ const Card = styled.div`
     line-height: 14px;
     margin-right: 50px;
   }
-  .transfer {
+
+  button {
+    border: 0;
     width: 85px;
     height: 85px;
     background-color: #00af54;
@@ -111,39 +120,10 @@ const Card = styled.div`
     justify-content: center;
     align-items: center;
     user-select: none;
+    transition: background 0.2s;
+
+    :hover {
+      background-color: ${darken(0.05, '#00af54')};
+    }
   }
 `;
-
-const HeaderMainPage: React.FC = () => {
-  return (
-    <Header>
-      <SearchContainer>
-        <SearchBox>
-          <div>
-            <BiSearch size="16px" color="#696969" />
-          </div>
-          <input placeholder="Buscar campeonato" />
-        </SearchBox>
-
-        <Icon>
-          <BiFilterAlt size="12px" color="#01582B" />
-        </Icon>
-      </SearchContainer>
-      <Card>
-        <div className="extract">
-          <span className="title">
-            Disponível <br />{' '}
-          </span>
-          <span className="money">$ 100,00 </span>
-          <hr />
-          <div className="extract-container">
-            <a href="#1">Ver extrato</a>
-          </div>
-        </div>
-        <div className="transfer">Transferir</div>
-      </Card>
-    </Header>
-  );
-};
-
-export default HeaderMainPage;
